+++ didact/package.json
@@ -16,6 +16,7 @@
     "react-scripts": "3.2.0",
     "redux": "^4.0.4",
     "redux-thunk": "^2.3.0",
+    "styled-components": "^4.4.0",
     "yup": "^0.27.0"
   },
   "scripts": {
