+++ didact/src/components/Register/register.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React from "react"
 import {Form, Field, withFormik} from "formik";
 import * as Yup from 'yup';
 import { registerAction, registerWithFacebook, registerWithGoogle } from '../../store/actions';
