+++ didact/src/components/mainPage/PageStyles.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const PageFlex = styled.div`
+    display: flex;
+`;
