+++ didact/src/store/actions/onboardingActions.js
@@ -13,44 +13,41 @@ export const VERIFY_FAILURE = 'VERIFY_FAILURE';
 
 
 export const loginAction = (history, form) => dispatch => {
-
-    dispatch({type: LOGIN_START})
-
-    axios.post(`https://didactlms-staging.herokuapp.com/api/auth/login`, form)
-        .then(res => 
-            {
-                localStorage.setItem("token", res.data.token)
-                dispatch({type: LOGIN_SUCCESS, payload: res})
-            })
+    dispatch({ type: LOGIN_START })
+    axios
+        .post(`https://didactlms-staging.herokuapp.com/api/auth/login`, form)
+        .then(res => {
+            localStorage.setItem("token", res.data.token)
+            dispatch({ type: LOGIN_SUCCESS, payload: res })
+        })
         .then(res => history.push("/dashboard"))
-        .catch(err => 
-            {
-                dispatch({type: LOGIN_FAILURE, payload: err})
-            })
+        .catch(err => {
+            dispatch({ type: LOGIN_FAILURE, payload: err })
+        })
 }
 
 export const registerAction = (history, form) => dispatch => {
     dispatch({ type: REGISTER_START });
     axios
-      .post("https://didactlms-staging.herokuapp.com/api/auth/register", form)
-      .then(res => {
-        dispatch({ type: REGISTER_SUCCESS, payload: res.data });
-        localStorage.setItem("token", res.data.token)
-      })
-      .then(res => history.push("/dashboard"))
-      .catch(err => dispatch({ type: REGISTER_FAILURE, payload: err }));
-  };
+        .post("https://didactlms-staging.herokuapp.com/api/auth/register", form)
+        .then(res => {
+            dispatch({ type: REGISTER_SUCCESS, payload: res.data });
+            localStorage.setItem("token", res.data.token)
+        })
+        .then(res => history.push("/dashboard"))
+        .catch(err => dispatch({ type: REGISTER_FAILURE, payload: err }));
+};
 
 export const verifyToken = (props) => dispatch => {
     console.log('props in action: ', props)
     console.log(localStorage.getItem('token'))
     const token = localStorage.getItem('token')
-    dispatch({type: VERIFY_START})
-    axios.post(`https://didactlms-staging.herokuapp.com/api/auth`, {'token': token})
-    .then(res => {
-        console.log(res)
-        dispatch({type: VERIFY_SUCCESS, payload: res.data})
-    })
-    .then(props.history.push('/dashboard'))
-    .catch(err => dispatch({type: VERIFY_FAILURE, payload: err}))
+    dispatch({ type: VERIFY_START })
+    axios.post(`https://didactlms-staging.herokuapp.com/api/auth`, { 'token': token })
+        .then(res => {
+            console.log(res)
+            dispatch({ type: VERIFY_SUCCESS, payload: res.data })
+        })
+        .then(props.history.push('/dashboard'))
+        .catch(err => dispatch({ type: VERIFY_FAILURE, payload: err }))
 }
