+++ didact/src/utils/Routes.jsx
@@ -2,16 +2,16 @@ import React from "react";
 import { Route} from "react-router-dom";
 import PrivateRoute from "./PrivateRoute";
 import FormikLoginForm from "../components/login/Login.jsx";
-import FormikRegisterForm from "../components/Register/register.jsx";
+import FormikRegisterForm from "../components/register/Register.jsx";
 import Dashboard from "../components/Dashboard/dashboard.jsx";
-import Header from '../components/headerAndFooter/Header'
 import Auth from '../auth/Auth';
+import MainPage from '../components/mainPage/MainPage'
 
 const Routes = () => {
     return (
         <>
         <Route path="/login" component={FormikLoginForm} />
-        <PrivateRoute exact path="/dashboard" component={Header} />
+        <PrivateRoute exact path="/dashboard" component={MainPage} />
         <Route path="/register" component={FormikRegisterForm} />
         <Route path='/auth' component={Auth} />
         </>
